requirements.txt
nicegui
pyyaml
requests
openai
nltk
sumy
colorama
jinja2
werkzeug
spacy
transformers
peewee
sentencepiece
jsonpickle